move windows and unix process monitors together into one file because a) otherwise there's nowhere to put common classes related to monitoring and b) there's lots of copied+pasted logic at the moment and would be nice to rationalize it; as part of this the undocumented (and untested, and I suspect unused) __main__ entry points for the windows/unix monitors are removed; people can easiyl write their own scripts that do this if they need it (probably customized), but it doesn't need to be part of pysys itself
